Allow overriding the passDepth via the config properties of cut()
If you have made a roughing pass, you don't need to do multiple passes
anymore to do a finish pass, as the material left is minimal. Instead,
it's best to do a single pass.
This patch allows you to specify a passDepth to the config properties of
cut() to override the one that is specified with the tool -- typically
you can set it to the material height so you can do a single pass.

Signed-off-by: Michel Pollet <buserror@gmail.com>